requirements.txt
asgiref==3.5.2
awscli==1.25.41
boto3==1.24.41
botocore==1.27.41
cffi==1.15.1
colorama==0.4.4
cryptography==37.0.4
distlib==0.3.4
Django==4.0.4
django-cors-headers==3.13.0
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
docutils==0.16
filelock==3.7.0
jmespath==1.0.1
Pillow==9.1.1
platformdirs==2.5.2
psycopg2==2.9.3
pyasn1==0.4.8
pycparser==2.21
PyJWT==1.7.1
python-dateutil==2.8.2
pytz==2022.1
PyYAML==5.4.1
rsa==4.7.2
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.11
virtualenv==20.14.1

